# new LinkedHashMap/LinkedHashSet implementation