… src/library/scala/collection/mutable/LinkedHashSet.scala
… src/library/scala/collection/mutable/OpenHashMap.scala
… test/benchmarks/src/main/scala/scala/collection/mutable/LinkedHashMapBenchmark2.scala
… test/junit/scala/collection/mutable/LinkedHashSetTest.scala