This PR is to fix scala/bug#11369. The old PR is closed, so create new one. The benchmark data will be added if all tests pass.