@liang3zy22 2022-11-22
All tests passed. So I have run some benchmark testing to get some performance data. I copied mutable/HashMapBenchmark2.scala and made the linkedhashmap/linkedhashset version. Please give comment to improve the benchmark program if needed.

```
The old linkedhashmap data:
Benchmark                           (size)  (stringsOnly)  Mode  Cnt         Score               Error   Units
LinkedHashMapBenchmark2.lhmBuild               10           true  avgt   40       158.949 ±       1.056  ns/op
LinkedHashMapBenchmark2.lhmBuild              100           true  avgt   40      2990.460 ±      84.548  ns/op
LinkedHashMapBenchmark2.lhmBuild             1000           true  avgt   40     35159.047 ±     354.869  ns/op
LinkedHashMapBenchmark2.lhmBuild            10000           true  avgt   40    553150.859 ±    2310.097  ns/op
LinkedHashMapBenchmark2.lhmFillColliding       10           true  avgt   40       236.445 ±       1.12   ns/op
LinkedHashMapBenchmark2.lhmFillColliding      100           true  avgt   40      7262.521 ±      42.028  ns/op
LinkedHashMapBenchmark2.lhmFillColliding     1000           true  avgt   40    281634.118 ±    844.832   ns/op
LinkedHashMapBenchmark2.lhmFillColliding    10000           true  avgt   40  22332434.136 ±  1621274.211 ns/op
LinkedHashMapBenchmark2.lhmFillRegular         10           true  avgt   40       124.998 ±       1.464  ns/op
LinkedHashMapBenchmark2.lhmFillRegular        100           true  avgt   40      2622.406 ±      31.329  ns/op
LinkedHashMapBenchmark2.lhmFillRegular       1000           true  avgt   40     32016.401 ±     133.902  ns/op
LinkedHashMapBenchmark2.lhmFillRegular      10000           true  avgt   40    508360.072 ±    1930.748  ns/op
LinkedHashMapBenchmark2.lhmGetExisting         10           true  avgt   40        84.727 ±       0.391  ns/op
LinkedHashMapBenchmark2.lhmGetExisting        100           true  avgt   40       883.492 ±      1.615   ns/op
LinkedHashMapBenchmark2.lhmGetExisting       1000           true  avgt   40     10246.926 ±     198.531  ns/op
LinkedHashMapBenchmark2.lhmGetExisting      10000           true  avgt   40    138553.664 ±     2615.972 ns/op
LinkedHashMapBenchmark2.lhmGetNone             10           true  avgt   40       814.774 ±       3.964  ns/op
LinkedHashMapBenchmark2.lhmGetNone            100           true  avgt   40       765.882 ±       5.509  ns/op
LinkedHashMapBenchmark2.lhmGetNone           1000           true  avgt   40      9492.448 ±      38.564  ns/op
LinkedHashMapBenchmark2.lhmGetNone          10000           true  avgt   40    177260.698 ±    490.294   ns/op
LinkedHashMapBenchmark2.lhmIterateEntries      10           true  avgt   40        70.186 ±       0.245  ns/op
LinkedHashMapBenchmark2.lhmIterateEntries     100           true  avgt   40      700.647 ±        2.379  ns/op
LinkedHashMapBenchmark2.lhmIterateEntries    1000           true  avgt   40     7091.794 ±      20.849   ns/op
LinkedHashMapBenchmark2.lhmIterateEntries   10000           true  avgt   40    71360.056 ±    214.509    ns/op
LinkedHashMapBenchmark2.lhmIterateKeys         10           true  avgt   40        37.512 ±       0.026  ns/op
LinkedHashMapBenchmark2.lhmIterateKeys        100           true  avgt   40       373.153 ±       0.342  ns/op
LinkedHashMapBenchmark2.lhmIterateKeys       1000           true  avgt   40      3729.026 ±      8.098   ns/op
LinkedHashMapBenchmark2.lhmIterateKeys      10000           true  avgt   40     37082.995 ±    184.722   ns/op

```