@liang3zy22 2022-11-23
The linkedhashset benchmarking data is as below:
```
Old linkedhashset benchmark data:
Benchmark                                 (size)  (stringsOnly)  Mode  Cnt         Score        Error  Units
LinkedHashMapBenchmark2.lhsBuild              10           true  avgt   40        142.391 ±     1.474  ns/op
LinkedHashMapBenchmark2.lhsBuild             100           true  avgt   40      2626.594 ±     46.261  ns/op
LinkedHashMapBenchmark2.lhsBuild            1000           true  avgt   40     31229.980 ±    261.343  ns/op
LinkedHashMapBenchmark2.lhsBuild           10000           true  avgt   40    515878.416 ±   3385.323  ns/op
LinkedHashMapBenchmark2.lhsContainsFalse      10           true  avgt   40       794.082 ±      0.987  ns/op
LinkedHashMapBenchmark2.lhsContainsFalse     100           true  avgt   40       743.354 ±      3.808  ns/op
LinkedHashMapBenchmark2.lhsContainsFalse    1000           true  avgt   40      9481.574 ±     42.296  ns/op
LinkedHashMapBenchmark2.lhsContainsFalse   10000           true  avgt   40     178484.693 ±   235.924  ns/op
LinkedHashMapBenchmark2.lhsContainsTrue       10           true  avgt   40        86.067 ±      0.181  ns/op
LinkedHashMapBenchmark2.lhsContainsTrue      100           true  avgt   40       873.667 ±      3.007  ns/op
LinkedHashMapBenchmark2.lhsContainsTrue     1000           true  avgt   40      9570.567 ±     66.776  ns/op
LinkedHashMapBenchmark2.lhsContainsTrue    10000           true  avgt   40     119420.781 ±   470.263  ns/op
LinkedHashMapBenchmark2.lhsFillColliding      10           true  avgt   40       243.952 ±        3.2  ns/op
LinkedHashMapBenchmark2.lhsFillColliding     100           true  avgt   40      7245.039 ±     54.031  ns/op
LinkedHashMapBenchmark2.lhsFillColliding    1000           true  avgt   40    286476.023 ±   1072.997  ns/op
LinkedHashMapBenchmark2.lhsFillColliding   10000           true  avgt   40  19971599.863 ±  97841.933  ns/op
LinkedHashMapBenchmark2.lhsFillRegular        10           true  avgt   40        136.601 ±     0.685  ns/op
LinkedHashMapBenchmark2.lhsFillRegular       100           true  avgt   40      2598.248 ±     38.631  ns/op
LinkedHashMapBenchmark2.lhsFillRegular      1000           true  avgt   40     30654.817 ±     63.219  ns/op
LinkedHashMapBenchmark2.lhsFillRegular     10000           true  avgt   40     523159.815 ±  1712.072  ns/op
LinkedHashMapBenchmark2.lhsIterate            10           true  avgt   40        37.178 ±      0.009  ns/op
LinkedHashMapBenchmark2.lhsIterate           100           true  avgt   40       367.921 ±      0.562  ns/op
LinkedHashMapBenchmark2.lhsIterate          1000           true  avgt   40      3696.711 ±     11.559  ns/op
LinkedHashMapBenchmark2.lhsIterate         10000           true  avgt   40     37111.388 ±     23.572  ns/op

New linkedhashset benchmark data:
Benchmark                                 (size)  (stringsOnly)  Mode  Cnt         Score        Error  Units
LinkedHashMapBenchmark2.lhsBuild              10           true  avgt   40        99.393 ±      1.299  ns/op
LinkedHashMapBenchmark2.lhsBuild             100           true  avgt   40      1964.677 ±      5.188  ns/op
LinkedHashMapBenchmark2.lhsBuild            1000           true  avgt   40     17275.521 ±     54.087  ns/op
LinkedHashMapBenchmark2.lhsBuild           10000           true  avgt   40    197580.457 ±    650.951  ns/op
LinkedHashMapBenchmark2.lhsContainsFalse      10           true  avgt   40       609.765 ±      0.774  ns/op
LinkedHashMapBenchmark2.lhsContainsFalse     100           true  avgt   40       592.982 ±      0.460  ns/op
LinkedHashMapBenchmark2.lhsContainsFalse    1000           true  avgt   40      6121.574 ±     25.423  ns/op
LinkedHashMapBenchmark2.lhsContainsFalse   10000           true  avgt   40     75333.319 ±    273.469  ns/op
LinkedHashMapBenchmark2.lhsContainsTrue       10           true  avgt   40        56.577 ±      0.155  ns/op
LinkedHashMapBenchmark2.lhsContainsTrue      100           true  avgt   40       576.741 ±      0.224  ns/op
LinkedHashMapBenchmark2.lhsContainsTrue     1000           true  avgt   40      6329.808 ±     28.641  ns/op
LinkedHashMapBenchmark2.lhsContainsTrue    10000           true  avgt   40     66922.317 ±    232.675  ns/op
LinkedHashMapBenchmark2.lhsFillColliding      10           true  avgt   40       216.958 ±      1.042  ns/op
LinkedHashMapBenchmark2.lhsFillColliding     100           true  avgt   40      6858.672 ±     52.100  ns/op
LinkedHashMapBenchmark2.lhsFillColliding    1000           true  avgt   40    360465.883 ±   7244.140  ns/op
LinkedHashMapBenchmark2.lhsFillColliding   10000           true  avgt   40  20239064.451 ± 141196.263  ns/op
LinkedHashMapBenchmark2.lhsFillRegular        10           true  avgt   40        87.968 ±      0.234  ns/op
LinkedHashMapBenchmark2.lhsFillRegular       100           true  avgt   40      1976.888 ±      7.802  ns/op
LinkedHashMapBenchmark2.lhsFillRegular      1000           true  avgt   40     17384.031 ±    174.278  ns/op
LinkedHashMapBenchmark2.lhsFillRegular     10000           true  avgt   40    199986.822 ±   1527.214  ns/op
LinkedHashMapBenchmark2.lhsIterate            10           true  avgt   40        37.459 ±      0.155  ns/op
LinkedHashMapBenchmark2.lhsIterate           100           true  avgt   40       371.253 ±      0.761  ns/op
LinkedHashMapBenchmark2.lhsIterate          1000           true  avgt   40      3718.015 ±     14.129  ns/op
LinkedHashMapBenchmark2.lhsIterate         10000           true  avgt   40     37219.866 ±     11.121  ns/op

```